Move Shape.kt and Type.kt theme tokens to commons
Shape.kt (410 lines):
- Move all shape/layout/modifier definitions to commons commonMain
- Inline TopBarSize (50.dp) to remove amethyst.ui.navigation dependency
- Update ChatTheme.kt to reference Shape.kt instead of duplicating values
- Replace amethyst Shape.kt with backward-compatible re-exports

Type.kt (132 lines):
- Move Typography, font size constants, and MarkdownTextStyle to commons commonMain
- Move DefaultHeadingStyle to commons jvmAndroidMain (depends on richtext lib)
- Change DefaultHeadingStyle from internal to public (cross-module access)
- Replace amethyst Type.kt with backward-compatible re-exports

Builds: iOS ✓ Android ✓ Desktop ✓
